refactor(stella-tui): split the transcript types and two test clusters out of the model fold
Closes #4217
